task2: review fixes
