+++ package.json
@@ -4,9 +4,8 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-        "build-prod": "webpack --mode production",
-        "build-dev": "webpack --mode development",
-        "start": "webpack-dev-server",
+        "build": "webpack --config webpack.prod.js",
+        "start": "webpack-dev-server --open --config webpack.dev.js",
         "test": "jest"
     },
     "repository": {
@@ -31,7 +30,8 @@
         "style-loader": "^0.23.1",
         "uglifyjs-webpack-plugin": "^2.1.2",
         "webpack": "^4.29.6",
-        "webpack-cli": "^3.3.0"
+        "webpack-cli": "^3.3.0",
+        "webpack-merge": "^4.2.1"
     },
     "dependencies": {
         "react": "^16.8.4",
