+++ webpack.dev.js
@@ -1,0 +1,13 @@
+const path = require('path');
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+    mode: 'development',
+
+    devtool: 'inline-source-map',
+    
+    devServer: {
+        contentBase: './dist'
+    }
+});
